+++ src/main/java/exceptions/EkudException.java
@@ -1,4 +1,5 @@
-package extensions;
+package exceptions;
+
 /**
  * Represents an exception thrown by the Ekud chatbot.
  */
