+++ src/main/java/exceptions/EkudIllegalArgException.java
@@ -1,4 +1,4 @@
-package extensions;
+package exceptions;
 /**
  * Represents illegal arguments by the user after giving a command to the chatbot.
  */
