+++ src/main/java/tasks/Event.java
@@ -1,4 +1,4 @@
-package extensions;
+package tasks;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
